Move knowledge wizard utils under knowledge component dir
